requirements.txt
appscript==1.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
et_xmlfile==2.0.0
idna==3.10
lxml==5.3.0
msal==1.31.0
psutil==6.1.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
